requirements.txt
Django==4.1.13
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.1
Requests==2.32.2
